# Provide previous app image to tekton for re-use.