--- a/internal/api/v1/application/stage.go
+++ b/internal/api/v1/application/stage.go
@@ -40,6 +40,7 @@ type stageParam struct {
 	S3ConnectionDetails s3manager.ConnectionDetails
 	Stage               models.StageRef
 	Username            string
+	PreviousStageID     string
 }
 
 // ImageURL returns the URL of the container image to be, using the
@@ -165,6 +166,15 @@ func (hc Controller) Stage(c *gin.Context) apierror.APIErrors {
 		UID:        app.GetUID(),
 	}
 
+	// Determine stage id of currently running deployment, fallback to itself when no such exists.
+	previousID, err := application.PreviousStageId(ctx, cluster, req.App)
+	if err != nil {
+		return apierror.InternalError(err, "failed to determine active application stage id")
+	}
+	if previousID == "" {
+		previousID = uid
+	}
+
 	registryPublicURL, err := getRegistryURL(ctx, cluster)
 	if err != nil {
 		return apierror.InternalError(err, "getting the Epinio registry public URL")
@@ -178,6 +188,7 @@ func (hc Controller) Stage(c *gin.Context) apierror.APIErrors {
 		RegistryURL:         registryPublicURL,
 		S3ConnectionDetails: s3ConnectionDetails,
 		Stage:               models.NewStage(uid),
+		PreviousStageID:     previousID,
 		Username:            username,
 	}
 
@@ -323,6 +334,10 @@ func newPipelineRun(app stageParam) *v1beta1.PipelineRun {
 		app.BlobUID,
 	}
 
+	// fake stage params of the previous to pull the old image url from.
+	previous := app
+	previous.Stage = models.NewStage(app.PreviousStageID)
+
 	return &v1beta1.PipelineRun{
 		ObjectMeta: metav1.ObjectMeta{
 			Name: app.Stage.ID,
@@ -341,6 +356,7 @@ func newPipelineRun(app stageParam) *v1beta1.PipelineRun {
 			PipelineRef:        &v1beta1.PipelineRef{Name: "staging-pipeline"},
 			Params: []v1beta1.Param{
 				{Name: "APP_IMAGE", Value: *str(app.ImageURL(app.RegistryURL))},
+				{Name: "PREVIOUS_IMAGE", Value: *str(previous.ImageURL(previous.RegistryURL))},
 				{Name: "BUILDER_IMAGE", Value: *str(app.BuilderImage)},
 				{Name: "ENV_VARS", Value: v1beta1.ArrayOrString{
 					Type:     v1beta1.ParamTypeArray,