--- a/internal/application/application.go
+++ b/internal/application/application.go
@@ -245,7 +245,11 @@ func deleteStagePVC(ctx context.Context, cluster *kubernetes.Cluster, appRef mod
 		PersistentVolumeClaims(deployments.TektonStagingNamespace).Delete(ctx, appRef.MakePVCName(), metav1.DeleteOptions{})
 }
 
-func PreviousStageId(ctx context.Context, cluster *kubernetes.Cluster, appRef models.AppRef) (string, error) {
+// PreviousStageID returns the stage ID of the previous build, if one exists. It returns an empty string otherwise.
+// This method relies on the presence of a workload to get the previous id. There is the case that staging has
+// happened, yet there is no workload. Ee.g. by calling the "staging" endpoint but not calling the "deploy"
+// endpoint. Since our client doesn't support that scenario, this method doesn't support it either.
+func PreviousStageID(ctx context.Context, cluster *kubernetes.Cluster, appRef models.AppRef) (string, error) {
 	wl := NewWorkload(cluster, appRef)
 
 	deployment, err := wl.Deployment(ctx)