fix #975 

consider it draft, with some more testing needed to confirm that the proper information is passed around.
it does not seem to break pushing, and on second push the `image not message` seen in any first push is gone.

the second commit extends the code a bit to attach the previous stage-id to the new pipelinerun resource, as a label.
confirmed with this that the existing code truly puts in the previous stage id for the run to reuse.