@andreas-kupries 2021-11-29
# Previous Image - Timing Changes

## Expectations

  - With the changes it is expected that the time to push and deploy the app is notably lower for second to fourth push, versus the first push.

  - Without the changes it is expected that the Time to push and deploy the app is roughly the same for all pushes, without the supplication of the previous image.

## Results

The expectations were met for the changed setup, and not met for unchanged setup.

Basically, both setups show a much faster push time for the second and further pushes. I suspect that other caching exist in the system, like for the sources themselves causing this.

In a direct comparison I can however note that with the change the second and further pushes are consistently about 13-17 seconds faster still (wall clock) than the pushes without.

### Comparison

With change to the left, without to the right.

|#   |Wall Time|User Time|Sys Time|   |Wall Time|User Time|Sys Time|
|---:|---:     |---:     |---:    |---|---:     |---:     |---:    |
|1   |4m46.132s|0m0.072s |0m0.050s|   |3m0.134s |0m0.069s |0m0.030s|
|2   |1m01.922s|0m0.062s |0m0.019s|   |1m14.896s|0m0.045s |0m0.033s|
|3   |1m01.921s|0m0.063s |0m0.020s|   |1m17.716s|0m0.063s |0m0.014s|
|4   |1m01.934s|0m0.055s |0m0.027s|   |1m18.722s|0m0.060s |0m0.017s|
|5   |0m57.917s|0m0.069s |0m0.009s|   |1m16.729s|0m0.060s |0m0.020s|

## Sources

  - Testing of the change was done with commit 32710af4c2d106558796709a779ae4377e73058e.

  - Testing without the change was done with commit 8227440c5ffabb4fffd02facdf32caee18f5aac9.

## Methodology

The command

  - `time epinio push --name fox --path assets/sample-app`

was executed 5 times on the same cluster and epinio installation.

__ATTENTION__: The 5 second wait for user confirmation at the beginning of each `push` was __not__ shortened by manual intervention. This was done to ensure consistent timing. We can subtract that period from the final time at the end.

## Testing

### Changed - 32710af4c2d106558796709a779ae4377e73058e

|#   |Wall Time|User Time|Sys Time|Notes|
|---:|---:     |---:     |---:    |---	|
|1   |4m46.132s|0m0.072s |0m0.050s|[1]	|
|2   |1m01.922s|0m0.062s |0m0.019s|[2]	|
|3   |1m01.921s|0m0.063s |0m0.020s|	|
|4   |1m01.934s|0m0.055s |0m0.027s|	|
|5   |0m57.917s|0m0.069s |0m0.009s|	|

[1]
```
step-create Previous image with name "epinio-registry.172.19.0.4.omg.howdoi.website/apps/workspace-fox:b225d20068db4503" not found 
```

[2]
```
step-create Restoring metadata for "paketo-buildpacks/ca-certificates:helper" from app image 
step-create Restoring metadata for "paketo-buildpacks/httpd:httpd" from app image 
step-create Restoring metadata for "paketo-buildpacks/php-dist:php" from app image 
step-create Restoring metadata for "paketo-buildpacks/php-web:php-web" from app image 
```

### Unchanged - 8227440c5ffabb4fffd02facdf32caee18f5aac9

|#   |Wall Time|User Time|Sys Time|Notes|
|---:|---:     |---:     |---:    |---	|
|1   |3m0.134s |0m0.069s |0m0.030s|[1]	|
|2   |1m14.896s|0m0.045s |0m0.033s|[1]	|
|3   |1m17.716s|0m0.063s |0m0.014s|[1]	|
|4   |1m18.722s|0m0.060s |0m0.017s|[1]	|
|5   |1m16.729s|0m0.060s |0m0.020s|[1]	|