@jimmykarily 2021-11-30
The reason all pushes are faster the second time is because of the `cache` volume: https://github.com/epinio/epinio/blob/main/internal/api/v1/application/stage.go#L357-L362 (search for `ensurePVC` too). What we are looking for here is additional improvement due to using layers from a previous build. It seems that something is happening indeed but I guess it all depends on the number of layer than can be re-used and such. I think we should set the previous image anyway and let the buildpack figure out how to make the best of it.