@andreas-kupries 2021-11-30
I fully agree. Without setting the previous image the system does not even have a chance for reuse. And as setting it is certainly not worse by what I have seen, I see no reason not to.